Move scores and store_output functions accordingly
